+++ src/java/org/apache/poi/hssf/record/formula/eval/UnaryMinusEval.java
@@ -32,7 +32,7 @@ public final class UnaryMinusEval implements Function {
 		// enforce singleton
 	}
 
-	public ValueEval evaluate(ValueEval[] args, int srcRow, short srcCol) {
+	public ValueEval evaluate(ValueEval[] args, int srcRow, int srcCol) {
 		if (args.length != 1) {
 			return ErrorEval.VALUE_INVALID;
 		}
