+++ src/java/org/apache/poi/hssf/record/formula/functions/Substitute.java
@@ -29,7 +29,7 @@ import org.apache.poi.hssf.record.formula.eval.ValueEval;
  */
 public final class Substitute extends TextFunction {
 
-	protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, short srcCellCol)
+	protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, int srcCellCol)
 			throws EvaluationException {
 		if (args.length < 3 || args.length > 4) {
 			return ErrorEval.VALUE_INVALID;
