+++ src/java/org/apache/poi/hssf/record/formula/functions/TextFunction.java
@@ -33,16 +33,16 @@ public abstract class TextFunction implements Function {
 
 	protected static final String EMPTY_STRING = "";
 
-	protected static final String evaluateStringArg(ValueEval eval, int srcRow, short srcCol) throws EvaluationException {
+	protected static final String evaluateStringArg(ValueEval eval, int srcRow, int srcCol) throws EvaluationException {
 		ValueEval ve = OperandResolver.getSingleValue(eval, srcRow, srcCol);
 		return OperandResolver.coerceValueToString(ve);
 	}
-	protected static final int evaluateIntArg(ValueEval arg, int srcCellRow, short srcCellCol) throws EvaluationException {
+	protected static final int evaluateIntArg(ValueEval arg, int srcCellRow, int srcCellCol) throws EvaluationException {
 		ValueEval ve = OperandResolver.getSingleValue(arg, srcCellRow, srcCellCol);
 		return OperandResolver.coerceValueToInt(ve);
 	}
 
-	public final ValueEval evaluate(ValueEval[] args, int srcCellRow, short srcCellCol) {
+	public final ValueEval evaluate(ValueEval[] args, int srcCellRow, int srcCellCol) {
 		try {
 			return evaluateFunc(args, srcCellRow, srcCellCol);
 		} catch (EvaluationException e) {
@@ -50,7 +50,7 @@ public abstract class TextFunction implements Function {
 		}
 	}
 
-	protected abstract ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, short srcCellCol) throws EvaluationException;
+	protected abstract ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, int srcCellCol) throws EvaluationException;
 
 	/* ---------------------------------------------------------------------- */
 
@@ -59,7 +59,7 @@ public abstract class TextFunction implements Function {
 		protected SingleArgTextFunc() {
 			// no fields to initialise
 		}
-		protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, short srcCellCol)
+		protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, int srcCellCol)
 				throws EvaluationException {
 			if (args.length != 1) {
 				return ErrorEval.VALUE_INVALID;
@@ -109,7 +109,7 @@ public abstract class TextFunction implements Function {
 	 */
 	public static final Function MID = new TextFunction() {
 
-		protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, short srcCellCol)
+		protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, int srcCellCol)
 				throws EvaluationException {
 			if (args.length != 3) {
 				return ErrorEval.VALUE_INVALID;
@@ -144,7 +144,7 @@ public abstract class TextFunction implements Function {
 		protected LeftRight(boolean isLeft) {
 			_isLeft = isLeft;
 		}
-		protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, short srcCellCol)
+		protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, int srcCellCol)
 				throws EvaluationException {
 			if (args.length != 2) {
 				return ErrorEval.VALUE_INVALID;
@@ -167,7 +167,7 @@ public abstract class TextFunction implements Function {
 
 	public static final Function CONCATENATE = new TextFunction() {
 
-		protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, short srcCellCol)
+		protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, int srcCellCol)
 				throws EvaluationException {
 			StringBuffer sb = new StringBuffer();
 			for (int i=0, iSize=args.length; i<iSize; i++) {
@@ -179,7 +179,7 @@ public abstract class TextFunction implements Function {
 
 	public static final Function EXACT = new TextFunction() {
 
-		protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, short srcCellCol)
+		protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, int srcCellCol)
 				throws EvaluationException {
 			if (args.length != 2) {
 				return ErrorEval.VALUE_INVALID;
